pyproject.toml
[tool.poetry]
name = "bankly"
version = "0.1.0"
description = ""
authors = ["amantur02 <amantur.ubaid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
fastapi = "^0.104.1"
pydantic-settings = "^2.0.3"
uvicorn = "^0.24.0.post1"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
alembic = "^1.12.1"
passlib = "^1.7.4"
aioredis = "2.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
